requirements/base.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
Unidecode==0.04.16
django-environ==0.4.0

amqp==1.4.4
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.16

carteblanche==1.0.0
celery==3.1.10
dj-static==0.0.6
django-activity-stream==0.4.5beta1
django-appconf==0.6
django-bootstrap-pagination==1.5.1
django-bower==4.8.1
django-celery==3.1.9
django-debug-toolbar==1.2.1
django-crispy-forms==1.6.0
django-email-extras==0.2
django-forms-builder==0.10
django-geoposition==0.2.0
django-json-field==0.5.5
django-jsonfield==0.9.12
django-object-log==0.7
-e git://github.com/hugotox/django-remote-forms.git@0b2b2327ac3251dcb041643288b66bd421671786#egg=django_remote_forms-master
django-storages==1.1.8
future==0.12.4

httplib2==0.9
ipython==1.2.0
jsonfield==0.9.20
kombu==3.0.14
logging==0.4.9.6
markdown2==2.2.0
oauthlib==0.6.1
psycopg2==2.5.4
pystache==0.3.1
python-dateutil==2.2
python-gnupg==0.3.6
pytz==2014.3
redis==2.9.1
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.1
six==1.6.1
sphinx-me==0.2.1
sqlparse==0.1.12
static3==0.6.1
twilio==3.6.15
wsgiref==0.1.2
xlwt==0.7.5

django-braces==1.8.1
